backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.28.1
requests==2.31.0
ollama==0.3.3
